More fix (ignore) more GCC compiler warnings

+++ include/mapnik/grid/grid_pixel.hpp
@@ -113,7 +113,7 @@ struct gray16
     self_type gradient(self_type c, double k) const
     {
         self_type ret;
-        calc_type ik = agg::uround(k * base_scale);
+        calc_type ik = agg::uround(k * static_cast<double>(base_scale));
         ret.v = value_type(calc_type(v) + (((calc_type(c.v) - v) * ik) >> base_shift));
         ret.a = value_type(calc_type(a) + (((calc_type(c.a) - a) * ik) >> base_shift));
         return ret;
@@ -241,7 +241,7 @@ struct gray32
     self_type gradient(self_type c, double k) const
     {
         self_type ret;
-        calc_type ik = agg::uround(k * base_scale);
+        calc_type ik = agg::uround(k * static_cast<double>(base_scale));
         ret.v = value_type(calc_type(v) + (((calc_type(c.v) - v) * ik) >> base_shift));
         ret.a = value_type(calc_type(a) + (((calc_type(c.a) - a) * ik) >> base_shift));
         return ret;
@@ -369,7 +369,7 @@ struct gray64
     self_type gradient(self_type c, double k) const
     {
         self_type ret;
-        calc_type ik = agg::uround(k * base_scale);
+        calc_type ik = agg::uround(k * static_cast<double>(base_scale));
         ret.v = value_type(calc_type(v) + (((calc_type(c.v) - v) * ik) >> base_shift));
         ret.a = value_type(calc_type(a) + (((calc_type(c.a) - a) * ik) >> base_shift));
         return ret;
